Combine webpack configs and include babelConfig
