+++ webpack.config.js
@@ -1,11 +1,13 @@
 "use strict";
 
 const webpack = require("webpack");
+const fs = require("fs");
 const path = require("path");
 const CopyPlugin = require("copy-webpack-plugin");
 const MiniCssExtractPlugin = require("mini-css-extract-plugin");
 const VueLoaderPlugin = require("vue-loader/lib/plugin");
 const Helper = require("./src/helper.js");
+const babelConfig = require("./babel.config.cjs");
 
 const isProduction = process.env.NODE_ENV === "production";
 const config = {
@@ -65,6 +67,7 @@ const config = {
 				include: [path.resolve(__dirname, "client")],
 				use: {
 					loader: "babel-loader",
+					options: babelConfig,
 				},
 			},
 		],
@@ -139,4 +142,52 @@ const config = {
 	],
 };
 
-module.exports = config;
+module.exports = (env, argv) => {
+	if (argv.mode === "development") {
+		const testFile = path.resolve(__dirname, "test/public/testclient.js");
+
+		if (fs.existsSync(testFile)) {
+			fs.unlinkSync(testFile);
+		}
+
+		config.target = "node";
+		config.devtool = "eval";
+		config.stats = "errors-only";
+		config.output.path = path.resolve(__dirname, "test/public");
+		config.entry = {
+			"testclient.js": [path.resolve(__dirname, "test/client/index.js")],
+		};
+
+		// Add the istanbul plugin to babel-loader options
+		for (const rule of config.module.rules) {
+			if (rule.use.loader === "babel-loader") {
+				rule.use.options.plugins = ["istanbul"];
+			}
+		}
+
+		// `optimization.splitChunks` is incompatible with a `target` of `node`. See:
+		// - https://github.com/zinserjan/mocha-webpack/issues/84
+		// - https://github.com/webpack/webpack/issues/6727#issuecomment-372589122
+		config.optimization.splitChunks = false;
+
+		// Disable plugins like copy files, it is not required
+		config.plugins = [
+			new VueLoaderPlugin(),
+
+			// Client tests that require Vue may end up requireing socket.io
+			new webpack.NormalModuleReplacementPlugin(
+				/js(\/|\\)socket\.js/,
+				path.resolve(__dirname, "scripts/noop.js")
+			),
+
+			// "Fixes" Critical dependency: the request of a dependency is an expression
+			new webpack.ContextReplacementPlugin(/vue-server-renderer$/),
+		];
+	}
+
+	if (argv.mode === "production") {
+		// ...
+	}
+
+	return config;
+};
